Change namespace name : Frontend -> GSharpCompiler
