+++ GSharpCompiler/Types/Program.cs
@@ -3,7 +3,7 @@ This class represents a G# program.
 A program is a list of statements.
 */
 
-namespace Frontend;
+namespace GSharpCompiler;
 
 class Program{
     public List<Stmt> Stmts {get; private set;}//How can i give access to the statements to read but not to write.
